Removing pagination logic from TestResults file
